Refactoring and review changes
