+++ packages/peer-test-automation/src/constants.js
@@ -1,0 +1,8 @@
+// TODO: accept these constants as cli arguments
+export const TOTAL_PEERS = 3
+export const TEST_RETRIES = 15;
+export const MIN_REQUIRED_CONNECTIONS = 4;
+
+export const TEST_INTERVAL = 8; // seconds
+export const WAIT_BEFORE_RETRY = 10; // seconds
+export const CHECK_CONNECTION_INTERVAL = 3; // seconds
